+++ CHANGES.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-- Nothing yet...
+- Format nordea personal account numbers as personal id numbers in getNumber()
 
 ## [1.2.0] - 2016-10-13